Include applications.commands scope to simplify auth
Discord lets us add desired scopes for a bot via the URL for authorization.
By adding applications.commands in this initial request, we no longer
need to additionally add the bot to the server after it first joins.

This should also mean that bot application's OAuth2 settings no longer
need to be set for either production or local dev.